refactor(plugin-events-manager): consolidate duplicate embed and button creation
- Move createEventEmbed() to shared utils.js module
- Move createEventButtons() to shared utils.js module
- Remove duplicate implementations from event-create.js
- Remove duplicate implementations from modal-handler.js
- Unified embed includes 'Starting Soon' indicator for events within 24 hours
- Reduced code duplication by ~80 lines

Addresses CodeRabbit feedback on duplicate code
